+++ tests/lib/Acceptance/SoftDeleteTest.php
@@ -140,7 +140,7 @@ class SoftDeleteTest extends TestCase
 
         $this->schema->repository()->delete((string) $post->getRouteKey());
 
-        $this->assertDeleted($post);
+        $this->assertModelMissing($post);
         $this->assertTrue($forceDeleted);
     }