+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file. This projec
 
 ### Changed
 
+- Added return types for internal methods, to remove deprecation warnings on PHP 8.1.
 - [#20](https://github.com/laravel-json-api/eloquent/pull/20) **BREAKING** To support PHP 8.1 we needed to rename the
   `ReadOnly` contract and trait. This is because PHP 8.1 introduced `readonly` as a reserved word. The following changes
   were made: